moved the rest and changed name of the file to CanOpenEDS.cs
